Applying Extract Class tecnique of Moving Characteristics Between Objects in ParamModel class
Code smells:
 * The ParamModel class has too much responsability

Affected class:
 * ParamModel.java

Solutions:
* Creating a new model named Command.java in package paramWrapper and extracting all methods related to Command in ParamModel
